routes OK pre rating

+++ Backend/routes/index.js
@@ -4,6 +4,10 @@ const router = express.Router()
 import bcrypt from 'bcrypt'
 import jwt from 'jsonwebtoken'
 
+import auth from '../middleware/auth.js'
+import multer from '../middleware/multer-config.js'
+import fs from 'fs'
+
 import Book from '../models/book.js'
 import User from '../models/user.js'
 
@@ -50,7 +54,7 @@ router.post('/auth/login', (req, res) => {
                         userId: user._id,
                         token: jwt.sign(
                             {userId: user._id},
-                            'RANDOM_TOKEN_SECRET',
+                            process.env.JWT_SECRET,
                             {expiresIn: '24h'}
                         )
                     })
@@ -60,14 +64,21 @@ router.post('/auth/login', (req, res) => {
         .catch(error => res.status(500).json({ error }))
 })
 
-router.post('/books', (req, res) => {
-    const book = new Book(...req.body)
+router.post('/books', auth, multer, (req, res) => {
+    const addedBook = JSON.parse(req.body.book)
+    delete addedBook._id
+    delete addedBook._userId
+    const book = new Book({
+        ...addedBook,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    })
     book.save()
         .then(() => res.status(201).json({ message: 'book saved !'}))
         .catch(error => res.status(400).json({ error }))
 })
 
-router.post('/books/:id/rating', (req, res) => {
+router.post('/books/:id/rating', auth, (req, res) => {
     delete req.body._id
     const book = new Book(...req.body)
     book.save()
@@ -75,16 +86,41 @@ router.post('/books/:id/rating', (req, res) => {
         .catch(error => res.status(400).json({ error }))
 })
 
-router.put('books/:id', (req, res) => {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'book updated with success !'}))
-        .catch(error => res.status(400).json({ error }))
+router.put('/books/:id', auth, multer, (req, res) => {
+    const modBook = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+
+    delete modBook._userId
+    Book.findOne({_id: req.params.id})
+        .then(book => {
+            if(book.userId !== req.auth.userId) {
+                res.status(401).json({message : "Not authorized"})
+            } else {
+                Book.updateOne({ _id: req.params.id }, { ...modBook, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'book updated with success !'}))
+                    .catch(error => res.status(400).json({ error }))
+            }
+        })
+        .catch(error => {res.status(400).json({error})})
 })
 
-router.delete('books/:id', (req, res) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'book deleted !'}))
-        .catch(error => res.status(400).json({ error }))
+router.delete('/books/:id', auth, (req, res) => {
+    Book.findOne({ _id: req.params.id })
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({message: "not authorized"})
+            } else {
+                const filename = book.imageUrl.split('/image/')[1]
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: 'book deleted !'}))
+                        .catch(error => res.status(400).json({ error }))
+                })
+            }
+        })
+    
 })
 
 export default router
